Moving all shared POJOs to common
