--- src/main/java/redox/datamodel/patientpush/Identifier_.java
@@ -1,57 +1,0 @@
-
-package redox.datamodel.patientpush;
-
-import java.util.HashMap;
-import java.util.Map;
-import com.fasterxml.jackson.annotation.JsonAnyGetter;
-import com.fasterxml.jackson.annotation.JsonAnySetter;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-    "ID",
-    "IDType"
-})
-public class Identifier_ {
-
-    @JsonProperty("ID")
-    private String iD;
-    @JsonProperty("IDType")
-    private String iDType;
-    @JsonIgnore
-    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
-
-    @JsonProperty("ID")
-    public String getID() {
-        return iD;
-    }
-
-    @JsonProperty("ID")
-    public void setID(String iD) {
-        this.iD = iD;
-    }
-
-    @JsonProperty("IDType")
-    public String getIDType() {
-        return iDType;
-    }
-
-    @JsonProperty("IDType")
-    public void setIDType(String iDType) {
-        this.iDType = iDType;
-    }
-
-    @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
-        return this.additionalProperties;
-    }
-
-    @JsonAnySetter
-    public void setAdditionalProperty(String name, Object value) {
-        this.additionalProperties.put(name, value);
-    }
-
-}
